* moved interfaces from entity management core to entity management and removed entity manamgent core * exposed dbset<t> on entity repository
