+++ EntityManagement.sln
@@ -11,8 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EntityManagement.Core", "EntityManagement.Core\EntityManagement.Core.csproj", "{CF983F8D-9276-4D3A-919C-B0CCF6857E7E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EntityManagement", "EntityManagement\EntityManagement.csproj", "{8F950344-CED0-4897-8EC2-4827BC7D3F17}"
 EndProject
 Global
@@ -21,10 +19,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CF983F8D-9276-4D3A-919C-B0CCF6857E7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CF983F8D-9276-4D3A-919C-B0CCF6857E7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CF983F8D-9276-4D3A-919C-B0CCF6857E7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CF983F8D-9276-4D3A-919C-B0CCF6857E7E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8F950344-CED0-4897-8EC2-4827BC7D3F17}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8F950344-CED0-4897-8EC2-4827BC7D3F17}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F950344-CED0-4897-8EC2-4827BC7D3F17}.Release|Any CPU.ActiveCfg = Release|Any CPU
