+++ EntityManagement/EntityManagement.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <Version>0.0.3</Version>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <Version>0.0.4</Version>
     <PackageProjectUrl>https://github.com/TheMagnificent11/entity-management</PackageProjectUrl>
     <Authors>Saji Weerasingham</Authors>
     <Company>NA</Company>
@@ -29,11 +29,10 @@
     <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="2.6.0">
       <ExcludeAssets>all</ExcludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="2.1.1" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.0.2">
       <ExcludeAssets>all</ExcludeAssets>
     </PackageReference>
-    <PackageReference Include="EntityManagement.Core" Version="0.0.2" />
   </ItemGroup>
 
 </Project>
